# apps/backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Game Zone Management System API",
    "description": "REST API for managing a gaming cafe: users, devices, plans, sessions, transactions, and inventory.",
    "contact": {
      "name": "Gaming Cafe Team"
    },
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/device-games": {
      "get": {
        "tags": [
          "device-games"
        ],
        "operationId": "list_device_games",
        "responses": {
          "200": {
            "description": "List device games",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceGamePaginationEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "device-games"
        ],
        "operationId": "create_device_game",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDeviceGameDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Create device game",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceGameEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/device-games/device/{device_id}": {
      "get": {
        "tags": [
          "device-games"
        ],
        "operationId": "list_device_games_by_device",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "description": "Device ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List device games by device",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceGamePaginationEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/device-games/game/{game_id}": {
      "get": {
        "tags": [
          "device-games"
        ],
        "operationId": "list_device_games_by_game",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "description": "Game ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List device games by game",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceGamePaginationEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/device-games/{id}": {
      "get": {
        "tags": [
          "device-games"
        ],
        "operationId": "get_device_game",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Device game ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get device game",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceGameEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "device-games"
        ],
        "operationId": "delete_device_game",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Device game ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Soft-deactivated (isActive=false)"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "device-games"
        ],
        "operationId": "update_device_game",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Device game ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDeviceGameDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Update device game",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceGameEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateDeviceGameDto": {
        "type": "object",
        "required": [
          "deviceId",
          "gameId"
        ],
        "properties": {
          "deviceId": {
            "type": "string",
            "format": "uuid"
          },
          "gameId": {
            "type": "string",
            "format": "uuid"
          },
          "installationDate": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "isActive": {
            "type": [
              "boolean",
              "null"
            ]
          }
        }
      },
...
      "DeviceGameDevice": {
        "type": "object",
        "required": [
          "id",
          "name",
          "deviceType"
        ],
        "properties": {
          "deviceType": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "location": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          }
        }
      },
      "DeviceGameEnvelope": {
        "type": "object",
        "required": [
          "success",
          "statusCode",
          "timestamp",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DeviceGameResponse"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "string"
          }
        }
      },
...
      "DeviceGameGame": {
        "type": "object",
        "required": [
          "id",
          "title"
        ],
        "properties": {
          "genre": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "DeviceGamePaginationEnvelope": {
        "type": "object",
        "required": [
          "success",
          "statusCode",
          "timestamp",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DeviceGamePaginationPage"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "string"
          }
        }
      },
      "DeviceGamePaginationPage": {
        "type": "object",
        "required": [
          "data",
          "total",
          "page",
          "limit",
          "totalPages"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DeviceGameResponse"
            }
          },
          "limit": {
            "type": "integer",
            "format": "int64"
          },
          "page": {
            "type": "integer",
            "format": "int64"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "DeviceGameResponse": {
        "type": "object",
        "required": [
          "id",
          "deviceId",
          "gameId",
          "installationDate",
          "isActive",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "device": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DeviceGameDevice"
              }
            ]
          },
          "deviceId": {
            "type": "string",
            "format": "uuid"
          },
          "game": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DeviceGameGame"
              }
            ]
          },
          "gameId": {
            "type": "string",
            "format": "uuid"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "installationDate": {
            "type": "string",
            "format": "date-time"
          },
          "isActive": {
            "type": "boolean"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "statusCode",
          "message",
          "error",
          "timestamp"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "timestamp": {
            "type": "string"
          }
        }
      },
...
      "UpdateDeviceGameDto": {
        "type": "object",
        "properties": {
          "installationDate": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "isActive": {
            "type": [
              "boolean",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT obtained from POST /auth/verify-otp"
      }
    }
  },
...
}